refactor(capture): rejoin the log, debug, core and read-projection clusters; ratchet baseline 67 -> 21
Final round of un-splitting, plus two files whose headers described each other's contents:

  extension_logs.go <- extension_log_redaction.go, extension_log_store.go
    redaction is part of the ingest path; every append already went through it.
  debug.go <- debug_logger.go, accessor_debug.go, http_debug_redaction.go
    accessor_debug.go's only non-hub edge was to http_debug_redaction.go — the "accessor"
    prefix put it with the counter accessors, the graph puts it here.
  capture.go <- capture-struct.go, capture_lifecycle.go, client_registry_setter.go
  helpers.go <- rate_limit.go, ttl.go       shared request plumbing
  accessors.go <- store_views.go, performance_store.go
    store_views.go reads like part of the store layer but depends *on* the accessors;
    PerformanceStore's only callers are the performance accessors now in this file.
  buffer_store.go <- buffer_clear.go        the Capture-side wrapper over BufferStore.clear*
  extension_state.go <- status.go
    status.go's header claims "/sync request-response envelopes"; the file actually holds
    ExtensionStatus and UpdateExtensionStatus, which write c.extensionState.* directly.
    settings.go's header meanwhile described status.go's contents — the two were swapped.
    settings.go keeps its own file (it is the only disk-I/O path here) with a fixed header.
  aliases.go <- lifecycle_observer.go, the debuglog re-exports from debug.go
  testhelpers.go -> testhelpers_test.go
    setupTestCapture/setupTestServer/setupToolHandler are used only by capture's own tests,
    yet the file imported "testing" into the production build. Moving it into the test build
    drops that from the shipped binary. (test_helpers.go stays a production file: its
    *ForTest helpers are called by tests in eight other packages.)

Baseline ratcheted 67 -> 21; wsconn (3 files) is under the limit and needs no entry.

Behavior preservation, mechanically:
  - normalized non-comment code-line multiset over internal/capture: 71 lines only in the
    old tree, 142 only in the new. Every one is accounted for by the wsconn method renames
    and types.* qualification, the A11yCache/SessionTracker deletion, or the six new
    mutation-verified wsconn tests. No unexplained line.
  - go doc -all exported-symbol diff: only those same two dead types.
