+++ internal/capture/accessors.go
@@ -206,3 +206,171 @@ func (c *Capture) GetAndDeleteBeforeSnapshot(correlationID string) (PerformanceS
 	defer c.mu.Unlock()
 	return c.perf.takeBeforeSnapshot(correlationID)
 }
+
+const (
+	maxPerformanceSnapshots = 100
+	maxBeforeSnapshots      = 50
+)
+
+// appendSnapshots stores snapshots by URL with oldest-entry eviction.
+func (s *PerformanceStore) appendSnapshots(snapshots []PerformanceSnapshot) {
+	for _, snapshot := range snapshots {
+		key := snapshot.URL
+		if key == "" {
+			continue
+		}
+
+		if _, exists := s.snapshots[key]; !exists {
+			s.snapshotOrder = append(s.snapshotOrder, key)
+		}
+		s.snapshots[key] = snapshot
+
+		for len(s.snapshots) > maxPerformanceSnapshots && len(s.snapshotOrder) > 0 {
+			oldestKey := s.snapshotOrder[0]
+			s.snapshotOrder = s.snapshotOrder[1:]
+			delete(s.snapshots, oldestKey)
+		}
+	}
+}
+
+// snapshotsList returns a detached list copy.
+func (s *PerformanceStore) snapshotsList() []PerformanceSnapshot {
+	if len(s.snapshots) == 0 {
+		return []PerformanceSnapshot{}
+	}
+	out := make([]PerformanceSnapshot, 0, len(s.snapshots))
+	for _, snapshot := range s.snapshots {
+		out = append(out, snapshot)
+	}
+	return out
+}
+
+// snapshotByURL returns one snapshot by URL key.
+func (s *PerformanceStore) snapshotByURL(url string) (PerformanceSnapshot, bool) {
+	snap, ok := s.snapshots[url]
+	return snap, ok
+}
+
+// storeBeforeSnapshot keeps a pre-action snapshot for perf diff correlation.
+func (s *PerformanceStore) storeBeforeSnapshot(correlationID string, snapshot PerformanceSnapshot) {
+	s.beforeSnapshots[correlationID] = snapshot
+	if len(s.beforeSnapshots) <= maxBeforeSnapshots {
+		return
+	}
+
+	// Preserve current semantics: remove an arbitrary key when over cap.
+	for key := range s.beforeSnapshots {
+		delete(s.beforeSnapshots, key)
+		break
+	}
+}
+
+// takeBeforeSnapshot retrieves and deletes a before-snapshot (consume-on-read).
+func (s *PerformanceStore) takeBeforeSnapshot(correlationID string) (PerformanceSnapshot, bool) {
+	snap, ok := s.beforeSnapshots[correlationID]
+	if ok {
+		delete(s.beforeSnapshots, correlationID)
+	}
+	return snap, ok
+}
+
+// clear resets performance snapshot/baseline/before-snapshot state.
+func (s *PerformanceStore) clear() {
+	s.snapshots = make(map[string]PerformanceSnapshot)
+	s.snapshotOrder = make([]string, 0)
+	s.baselines = make(map[string]PerformanceBaseline)
+	s.baselineOrder = make([]string, 0)
+	s.beforeSnapshots = make(map[string]PerformanceSnapshot)
+}
+
+// EventBufferStore exposes high-volume event buffers through read-only snapshots.
+type EventBufferStore interface {
+	NetworkBodies() []NetworkBody
+	WebSocketEvents() []WebSocketEvent
+	EnhancedActions() []EnhancedAction
+}
+
+// NetworkWaterfallStore exposes network-waterfall snapshots.
+type NetworkWaterfallStore interface {
+	Entries() []NetworkWaterfallEntry
+	Count() int
+}
+
+// ExtensionLogStore exposes extension log snapshots.
+type ExtensionLogStore interface {
+	Entries() []ExtensionLog
+}
+
+// PerformanceSnapshotStore exposes performance snapshots keyed by URL.
+type PerformanceSnapshotStore interface {
+	Snapshots() []PerformanceSnapshot
+	SnapshotByURL(url string) (PerformanceSnapshot, bool)
+}
+
+type eventBufferView struct {
+	capture *Capture
+}
+
+func (v eventBufferView) NetworkBodies() []NetworkBody {
+	return v.capture.GetNetworkBodies()
+}
+
+func (v eventBufferView) WebSocketEvents() []WebSocketEvent {
+	return v.capture.GetAllWebSocketEvents()
+}
+
+func (v eventBufferView) EnhancedActions() []EnhancedAction {
+	return v.capture.GetAllEnhancedActions()
+}
+
+type networkWaterfallView struct {
+	capture *Capture
+}
+
+func (v networkWaterfallView) Entries() []NetworkWaterfallEntry {
+	return v.capture.GetNetworkWaterfallEntries()
+}
+
+func (v networkWaterfallView) Count() int {
+	return v.capture.GetNetworkWaterfallCount()
+}
+
+type extensionLogView struct {
+	capture *Capture
+}
+
+func (v extensionLogView) Entries() []ExtensionLog {
+	return v.capture.GetExtensionLogs()
+}
+
+type performanceSnapshotView struct {
+	capture *Capture
+}
+
+func (v performanceSnapshotView) Snapshots() []PerformanceSnapshot {
+	return v.capture.GetPerformanceSnapshots()
+}
+
+func (v performanceSnapshotView) SnapshotByURL(url string) (PerformanceSnapshot, bool) {
+	return v.capture.GetPerformanceSnapshotByURL(url)
+}
+
+// EventBuffers returns a read-only sub-store view for network/websocket/action buffers.
+func (c *Capture) EventBuffers() EventBufferStore {
+	return eventBufferView{capture: c}
+}
+
+// NetworkWaterfallStore returns a read-only sub-store view for waterfall entries.
+func (c *Capture) NetworkWaterfallStore() NetworkWaterfallStore {
+	return networkWaterfallView{capture: c}
+}
+
+// ExtensionLogStore returns a read-only sub-store view for extension logs.
+func (c *Capture) ExtensionLogStore() ExtensionLogStore {
+	return extensionLogView{capture: c}
+}
+
+// PerformanceSnapshotStore returns a read-only sub-store view for performance snapshots.
+func (c *Capture) PerformanceSnapshotStore() PerformanceSnapshotStore {
+	return performanceSnapshotView{capture: c}
+}
