+++ internal/capture/extension_logs.go
@@ -1,10 +1,12 @@
-// Purpose: Implements ingestion and retrieval of extension-internal debug logs captured via sync channels.
-// Why: Enables debugging of extension runtime behavior that page-level console capture cannot observe.
+// Purpose: Implements extension-internal log ingestion, redaction, ring-buffer storage and retrieval.
+// Why: Redaction is part of the ingest path — every append already went through it.
 // Docs: docs/features/feature/backend-log-streaming/index.md
+// Docs: docs/features/feature/redaction-patterns/index.md
 
 package capture
 
 import (
+	"encoding/json"
 	"time"
 )
 
@@ -48,4 +50,79 @@ func (c *Capture) GetExtensionLogs() []ExtensionLog {
 	c.mu.RLock()
 	defer c.mu.RUnlock()
 	return c.extensionLogs.snapshot()
+}
+
+// redactExtensionLog scrubs sensitive data from extension log fields before storage.
+func (c *Capture) redactExtensionLog(log ExtensionLog) ExtensionLog {
+	if c.logRedactor == nil {
+		return log
+	}
+
+	log.Message = c.logRedactor.Redact(log.Message)
+	log.Source = c.logRedactor.Redact(log.Source)
+	log.Category = c.logRedactor.Redact(log.Category)
+	if len(log.Data) > 0 {
+		log.Data = c.redactExtensionLogData(log.Data)
+	}
+	return log
+}
+
+func (c *Capture) redactExtensionLogData(data json.RawMessage) json.RawMessage {
+	var value any
+	if err := json.Unmarshal(data, &value); err != nil {
+		return json.RawMessage(c.logRedactor.Redact(string(data)))
+	}
+
+	redacted := redactJSONValue(value, c.logRedactor.Redact)
+	output, err := json.Marshal(redacted)
+	if err != nil {
+		return json.RawMessage(c.logRedactor.Redact(string(data)))
+	}
+	return output
+}
+
+func redactJSONValue(value any, redactFn func(string) string) any {
+	switch typed := value.(type) {
+	case map[string]any:
+		for key, child := range typed {
+			typed[key] = redactJSONValue(child, redactFn)
+		}
+		return typed
+	case []any:
+		for i, child := range typed {
+			typed[i] = redactJSONValue(child, redactFn)
+		}
+		return typed
+	case string:
+		return redactFn(typed)
+	default:
+		return value
+	}
+}
+
+// append adds one extension log entry and applies amortized eviction.
+func (b *ExtensionLogBuffer) append(log ExtensionLog) {
+	b.logs = append(b.logs, log)
+	evictionThreshold := MaxExtensionLogs + MaxExtensionLogs/2
+	if len(b.logs) <= evictionThreshold {
+		return
+	}
+
+	kept := make([]ExtensionLog, MaxExtensionLogs)
+	copy(kept, b.logs[len(b.logs)-MaxExtensionLogs:])
+	b.logs = kept
+}
+
+// snapshot returns a detached copy of the buffer contents.
+func (b *ExtensionLogBuffer) snapshot() []ExtensionLog {
+	out := make([]ExtensionLog, len(b.logs))
+	copy(out, b.logs)
+	return out
+}
+
+// clear removes all buffered logs and returns removed count.
+func (b *ExtensionLogBuffer) clear() int {
+	count := len(b.logs)
+	b.logs = make([]ExtensionLog, 0)
+	return count
 }
