--- internal/capture/http_debug_redaction.go
@@ -1,32 +1,0 @@
-// Purpose: Implements redaction of HTTP debug entry fields before persistence/output.
-// Why: Prevents sensitive request/response values from leaking through debug diagnostics.
-// Docs: docs/features/feature/redaction-patterns/index.md
-
-package capture
-
-// redactHTTPDebugEntry scrubs sensitive data from HTTP debug entry fields before storage.
-func (c *Capture) redactHTTPDebugEntry(entry HTTPDebugEntry) HTTPDebugEntry {
-	if c.logRedactor == nil {
-		return entry
-	}
-
-	if len(entry.Headers) > 0 {
-		redactedHeaders := make(map[string]string, len(entry.Headers))
-		for key, value := range entry.Headers {
-			redactedHeaders[key] = c.logRedactor.Redact(value)
-		}
-		entry.Headers = redactedHeaders
-	}
-
-	if entry.RequestBody != "" {
-		entry.RequestBody = c.logRedactor.Redact(entry.RequestBody)
-	}
-	if entry.ResponseBody != "" {
-		entry.ResponseBody = c.logRedactor.Redact(entry.ResponseBody)
-	}
-	if entry.Error != "" {
-		entry.Error = c.logRedactor.Redact(entry.Error)
-	}
-
-	return entry
-}
