+++ internal/capture/testhelpers_test.go
